+++ packages/annotation-toolkit/release-notes.md
@@ -1,3 +1,13 @@
+# vNext
+
+- **NEW** `<VideoPlayer />` will now read and execute callback fns from the data prop `renderSteps`. These steps can be defined by client code. The expected format of `renderSteps` is an object as such:
+
+```
+{
+  'render-key-id': (ctx, now, metadata, canvasRef) => {...}
+}
+```
+
 # v1.1.2-beta
 - **NEW** Adds in the `<MovableRect />` component with built-in linear interpolation between keyframes. Usage:
    ```js